#2035: Move /tags/ and /subjects/ to published controller
