+++ src/test/scala/no/ndla/conceptapi/service/ReadServiceTest.scala
@@ -17,7 +17,7 @@ class ReadServiceTest extends UnitSuite with TestEnvironment {
 
   test("Checks that filter by language works as it should") {
 
-    when(draftConceptRepository.everyTagFromEveryConcept).thenReturn(
+    when(publishedConceptRepository.everyTagFromEveryConcept).thenReturn(
       List(
         List(
           domain.ConceptTags(Seq("konge", "bror"), "nb"),
